requirements.txt
biopython==1.84
iniconfig==2.0.0
mmh3==5.0.1
numpy==2.1.3
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
